+++ src/temperatureAccessory.ts
@@ -53,9 +53,20 @@ export class TemperatureAccessory {
    */
   get(callback: CharacteristicGetCallback) {
     const temperature = this.platform.spa.getCurrentTemp();
-    this.platform.log.debug('Get Temperature Characteristic ->', temperature);
 
-    callback(null, temperature);
+    // Seems as if Homekit interprets null as something simply to be ignored, hence Homekit
+    // just uses the previous known value.
+    const val = (temperature == undefined ? null : temperature);
+    this.platform.log.debug('Get Temperature Characteristic ->', val);
+
+    callback(null, val);
+  }
+
+  // If Spa state has changed, for example using manual controls on the spa, then we must update Homekit.
+  updateCharacteristics() {
+    const temperature = this.platform.spa.getCurrentTemp();
+    const val = (temperature == undefined ? null : temperature!);
+    this.service.getCharacteristic(this.platform.Characteristic.CurrentTemperature).updateValue(val);
   }
 
 }